reverted back to a single cosi client (sry @Liamsi)
